finshing up getting file by hash..

+++ com/myster/hash/FileHash.java
@@ -2,7 +2,7 @@ package com.myster.hash;
 
 public abstract class FileHash {
 	public abstract byte[] getBytes() ;
-	public abstract int getHashLength() ;
+	public abstract short getHashLength() ;
 	public abstract String getHashName();
 	//toString <- should be hex value
 }
